+++ src/ccac/__init__.py
@@ -3,5 +3,4 @@
 from .validator import validate_document, validate_run_directory
 
 __all__ = ["validate_document", "validate_run_directory"]
-__version__ = "0.1.0"
-
+__version__ = "0.2.0"